+++ src/app/services/form.service.ts
@@ -57,6 +57,7 @@ export class FormService {
   snomedUser = false;
   _lformsVersion = '';
   _lformsErrorMessage = null;
+  _windowOpenerUrl: string = null;
 
   fetchService = inject(FetchService);
   formLevelExtensionService = inject(ExtensionsService);
@@ -152,6 +153,14 @@ export class FormService {
    */
   getFormLevelSchema() {
     return this.flSchema;
+  }
+
+  get windowOpenerUrl(): string {
+    return this._windowOpenerUrl;
+  }
+
+  set windowOpenerUrl(url: string) {
+    this._windowOpenerUrl = url;
   }
 
   get lformsVersion(): string {
@@ -444,7 +453,9 @@ export class FormService {
    */
   convertFromR4(fhirQ: fhir.Questionnaire, version: string): fhir.Questionnaire {
     let ret = fhirQ;
-    if(version !== 'R4') {
+    if (version === 'LHC-Forms') {
+      ret = LForms.Util.convertFHIRQuestionnaireToLForms(fhirQ);
+    } else if (version !== 'R4') {
       ret = LForms.Util.getFormFHIRData(fhirQ.resourceType, version,
         LForms.Util.convertFHIRQuestionnaireToLForms(fhirQ));
     }
@@ -503,10 +514,8 @@ export class FormService {
    * @param data - Data to post.
    */
   notifyWindowOpener(data: any) {
-    let openerUrl = window.opener?.location?.href;
-    if(openerUrl) {
-      window.opener.postMessage(data, openerUrl);
-      console.log(`${data.type} data posted to ${openerUrl}`);
+    if(this._windowOpenerUrl) {
+      window.opener.postMessage(data, this._windowOpenerUrl);
     }
   }