+++ tests/window-open-test.spec.ts
@@ -1,7 +1,22 @@
 import {test, expect, Page} from '@playwright/test';
 
 import {MainPO} from "./po/main-po";
-test.describe('Open form builder in a new window', () => {
+
+test.describe('Window opener notice', async () => {
+
+  test('should not exist if not opened by window.open() call', async ({page}) => {
+    await page.goto('/');
+    const mainPO = new MainPO(page);
+    await mainPO.loadFLPage();
+    // Note: Trying to assert non-existent element that it should NOT exist.
+    // However, after loading main page the element should have been attached,
+    // if it was opened by window.open(). So, negative assertion has still some
+    // testing value.
+    await expect(page.getByText(MainPO.windowOpenerNotice)).not.toBeAttached();
+  });
+});
+
+test.describe('Open form builder in a new window', async () => {
   let mainPO: MainPO;
 
 
@@ -23,6 +38,7 @@ test.describe('Open form builder in a new window', () => {
     mainPO = new MainPO(newPage);
     mainPO.clearSession();
     await mainPO.loadFLPage();
+    await expect(mainPO.page.getByText(MainPO.windowOpenerNotice)).toBeVisible();
   });
 
   test('should open form builder in a new window', async ({page}) => {